Refactor project structure and update configuration files for v16 ready
- Enhanced .gitignore to include additional file types and directories.
- Removed unused desktop and docs configuration files.
- Updated __init__.py to import taxes_and_totals correctly.
- Added new overrides for taxes_and_totals functionality.
- Introduced .editorconfig and .eslintrc for consistent coding standards.
- Added .pre-commit-config.yaml for pre-commit hooks.
- Created pyproject.toml for project metadata and dependencies.
