+++ navari_hotel_addon/__init__.py
@@ -1,3 +1,5 @@
+__version__ = "0.0.1"
 
-__version__ = '0.0.1'
+import navari_hotel_addon.navari_hotel_addon.overrides.taxes_and_totals as taxes_and_totals
 
+__all__ = ["taxes_and_totals"]
