--- navari_hotel_addon/config/desktop.py
@@ -1,12 +1,0 @@
-from frappe import _
-
-def get_data():
-	return [
-		{
-			"module_name": "Navari Hotel Addon",
-			"color": "grey",
-			"icon": "octicon octicon-file-directory",
-			"type": "module",
-			"label": _("Navari Hotel Addon")
-		}
-	]
