--- navari_hotel_addon/config/docs.py
@@ -1,11 +1,0 @@
-"""
-Configuration for docs
-"""
-
-# source_link = "https://github.com/[org_name]/navari_hotel_addon"
-# docs_base_url = "https://[org_name].github.io/navari_hotel_addon"
-# headline = "App that does everything"
-# sub_heading = "Yes, you got that right the first time, everything"
-
-def get_context(context):
-	context.brand_html = "Navari Hotel Addon"
